feat: implement basic JSON serialization for DynamicOptic
