+++ schema/shared/src/main/scala-3/zio/blocks/schema/migration/DynamicMigration.scala
@@ -1,9 +1,8 @@
 package zio.blocks.schema.migration
 
 import zio.blocks.schema.DynamicValue
-import zio.json.{JsonDecoder, JsonEncoder}
 
-final case class DynamicMigration(actions: Vector[MigrationAction]) derives JsonEncoder, JsonDecoder {
+final case class DynamicMigration(actions: Vector[MigrationAction]) {
   def apply(value: DynamicValue): Either[MigrationError, DynamicValue] =
     actions.foldLeft[Either[MigrationError, DynamicValue]](Right(value)) { (acc, action) =>
       acc.flatMap(action.apply)
